refactor: 위젯 색상 정의 위치 정리
- 추천 주제와 스트릭의 색상 구성을 전용 파일로 분리\n- 위젯 렌더링 파일에서 테마 색상 의존성 제거
